requirements.txt
flet>=0.21.0
mysql-connector-python>=8.0.0
